Update aws-java-sdk version and change hadoop-aws version to reflect this
Updating the hadoop-aws version to a non-standard version entailed a
series of difficulties due to transitive dependencies depending on the
$project.version. To avoid solving all the conflicts one by one, the
test module and its dependencies were removed. Note that prior to this
all unit tests were executed with the aws-java-sdk update to ensure no
functionality was broken. No further changes should be done to this
branch. Otherwise, the test module should be restored and executed.